[backend/frontend] fix errors + refacto (#7328)
