+++ opencti-platform/opencti-front/lang/front/ko.json
@@ -3866,6 +3866,7 @@
   "Used TTPs": "사용된 TTPs",
   "USER": "사용자",
   "User": "사용자",
+  "User cannot be created automatically for this connector since no default group has been defined.": "기본 그룹이 정의되어 있지 않으므로 이 커넥터에 대해 사용자를 자동으로 만들 수 없습니다.",
   "User cannot be created automatically for this feed since no default group has been defined.": "기본 그룹이 정의되지 않았으므로 이 피드에 대해 사용자를 자동으로 만들 수 없습니다.",
   "User cannot be created automatically for this feed since no default group has been defined. ": "기본 그룹이 정의되어 있지 않으므로 이 피드에 대해 사용자를 자동으로 만들 수 없습니다.",
   "User experience": "사용자 경험",
