+++ opencti-platform/opencti-front/src/private/components/data/IngestionCatalog.tsx
@@ -47,6 +47,7 @@ const IngestionCatalogComponent = ({
         {catalogs.map((catalog) => {
           return catalog.contracts.length > 0 && (
             <ListCardsContent
+              key={catalog.id}
               hasMore={() => false}
               isLoading={() => false}
               dataList={catalog.contracts}
